% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_sncm_fit}
\alias{plot_sncm_fit}
\title{plots the fit of an OTU table to \code{sncm}}
\usage{
plot_sncm_fit(spp_out = spp_out, fill = NULL, title = NULL)
}
\arguments{
\item{spp_out}{the output from \code{\link{fit_sncm}}}

\item{fill}{(optional) can either be set to fit_class to color by prediction (default), or by a taxonomic level
available in \code{\link[phyloseq]{rank_names}}}

\item{title}{(optional) the title of the plot.}
}
\value{
This function returns a plot of the fit of the OTU table to sncm
}
\description{
This funtion plots the output from \code{\link{fit_sncm}}
}
\seealso{
\code{\link{fit_sncm}}
}
